Initial build of process linker
This is initial commit in which we build an abstraction layer that masks actual implementation of communication for end users of `Linker` interface. `Linker` is basically a class which is used for creating connections between multiple processes connected in a given topology. They are identified simply by their integer ID so the end user doesn't need to know their hostnames or ports of their listener sockets. This makes communication simpler so we can send or receive messages by just providing ids of given processes. We assume that all messages are divided in 4 parts: source ID, destination ID, tag, message body.

We test this implementation in `LinkerTester` class.